ssh_key_gssapirenewalforcesrekey_test.go
package ssh_config

import "testing"

// GSSAPIRenewalForcesRekey
//         If set to “yes” then renewal of the client's GSSAPI credentials
//         will force the rekeying of the ssh connection. With a compati‐
//         ble server, this can delegate the renewed credentials to a ses‐
//         sion on the server.  The default is “no”.
//

func TestGSSAPIRenewalForcesRekeyInitFunc(t *testing.T) {
	defer func() {
		if err := recover(); err != nil {
			t.Errorf("panic in test is not acceptable")
		}
	}()
	f, ok := mapInit[GSSAPIRenewalForcesRekey]
	if !ok {
		t.Errorf("Cannot found init-function")
	}
	if f == nil {
		t.Errorf("Init-function is nil")
	}
	if f() == "" {
		// TODO
		t.Errorf("Not acceptable empty init value")
	}
}

func TestGSSAPIRenewalForcesRekeyValidFunc(t *testing.T) {
	defer func() {
		if err := recover(); err != nil {
			t.Errorf("panic in test is not acceptable")
		}
	}()
	f, ok := mapValid[GSSAPIRenewalForcesRekey]
	if !ok {
		t.Errorf("Cannot found valid-function")
	}
	if f == nil {
		t.Errorf("Valid-function is nil")
	}
	if !f("") {
		// TODO
		t.Errorf("Not acceptable empty valid-value")
	}
}

func TestGSSAPIRenewalForcesRekeyParseFunc(t *testing.T) {
	defer func() {
		if err := recover(); err != nil {
			t.Errorf("panic in test is not acceptable")
		}
	}()
	f, ok := mapParse[GSSAPIRenewalForcesRekey]
	if !ok {
		t.Errorf("Cannot found parse-function")
	}
	if f == nil {
		t.Errorf("Parse-function is nil")
	}
	// TODO
}
